zzip: cast pointer to "size_t", not "long"
A long is smaller than a pointer on WIN64.

+++ src/zzip/zip.c
@@ -388,8 +388,8 @@ static _zzip_inline char *
 __zzip_aligned4(char *p)
 {
 #define aligned4   __zzip_aligned4
-    p += ((long) p) & 1;        /* warnings about truncation of a "pointer" */
-    p += ((long) p) & 2;        /* to a "long int" may be safely ignored :) */
+    p += ((size_t) p) & 1; /* warnings about truncation of a "pointer" */
+    p += ((size_t) p) & 2; /* to a "long int" may be safely ignored :) */
     return p;
 }
